+++ lib/ruby3_backward_compatibility/compatibility/fixnum.rb
@@ -1,0 +1,3 @@
+unless defined?(::Fixnum)
+  Fixnum = Integer
+end